fix: 웹 지원자 카드 컴포넌트 분리
